requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
google-generativeai
pinecone-client
python-dotenv